Addressed bug #573 in part.
The current policy is as follows.  If the command string begins with \,
or is quoted with explicit '...' or "...", then alias-expansion does not
occur.  Builtin quotes $'...' are not considered.

   Shell. +=
       foo(argv) =
           echo Foo
   
   osh>foo
   Foo
   osh>\foo
   foo: command not found
   osh>"foo"
   foo: command not found
   osh>FOO = $'foo'
   osh>$(FOO)
   Foo
   osh>MOO = $'"foo"'
   osh>$(MOO)
   foo: command not found

One can argue that this is reasonably easy to understand.

The implementation of Aleksey's alternate theme (where $'foo' is not
alias-expanded, but 'foo' is) is hard.  This is because the following
two values are indistinguishable to omake.

   foo == $'foo'


git-svn-id: http://svn.metaprl.org/svnroot/mojave/omake-branches/0.9.8.x@9061 5109f7fa-8300-0410-942d-a7a480986ffa