Pass rule options as a Map, not an array of length-2 arrays (which is not a sensible value anyway).
Also, we weren't expanding :value: dependencies in
Omake_rule.eval_rule.  Suppose we had a rule like the
following.

   a: b :value: $X :exists: foo
      ...

The dependency is parsed as <sequence $`X " ">.  This is
a primitive value, so the $`X was never evaluated (and
so changing the value of X had no effect).

To do it right, flatten the values with values_of_value,
then take the primitive values.  You will see that the
value counts are now much larger.


git-svn-id: http://svn.metaprl.org/svnroot/mojave/omake-branches/0.9.8.x@9343 5109f7fa-8300-0410-942d-a7a480986ffa