More minor cleanup, addressing issue #516.
This adds the phony targets
   .BUILD_BEGIN
   .BUILD_SUCCESS
   .BUILD_FAILURE

.BUILD_BEGIN is built before anything else in your project.
One of .BUILD_SUCCESS or .BUILD_FAILURE is built when the
build for your project terminates.

These are normal targets/rules.  If you want to use them, you
should probably add them to your ~/.omakerc, rather than adding
them directly to each of the projects you work on.


git-svn-id: http://svn.metaprl.org/svnroot/mojave/omake-branches/0.9.8.x@9390 5109f7fa-8300-0410-942d-a7a480986ffa