This changes the way that .BUILD_* targets are built.
They are still called .BUILD_BEGIN, .BUILD_SUCCESS, and
.BUILD_FAILURE.  However, instead of mixing these targets
into the main build graph, we build them with a fresh
set of worklist queues.

Note, FAM events during the .BUILD_* phases are queued
and processed once the phase is done.

Also, any dependencies are pulled off of the main worklist
during a .BUILD phase.  I think this should be changed.


git-svn-id: http://svn.metaprl.org/svnroot/mojave/omake-branches/0.9.8.x@9441 5109f7fa-8300-0410-942d-a7a480986ffa