Made the omake_options type fully abstract.
This is needed to allow some "gap" between how the options are set and how the
options are used.  Currently some options have "side-effects" (e.g. -p also
enables -k and --no-s wouls also enable --print-status and --print-exit). The
way this is currently implemented is somewhat problematic - for examplle, if
one uses an explicit --no-k -p, the -p would still override the --no-k. With
the omake_options type being abstract it should be easier to have options that
oly affect the _default_ values of some other options, but do not override the
value if it was explicitly given by the user.

See also bug 172.


git-svn-id: http://svn.metaprl.org/svnroot/mojave/omake-branches/0.9.8.x@9744 5109f7fa-8300-0410-942d-a7a480986ffa