Minor changes for readline.c to compile on Win32 (untested).
Sometime, the "warnings as errors" flag was added as a Win32
C compilation flag.  This seems borken, since the OCaml manual
suggest to ignore the errors.  In the meantime, I added some
   (void) caml__dummy_request;
lines to avoid the error, but this is nonportable.  We should
probably turn off "warnings as errors" in the few selected
cases where this happens.


git-svn-id: http://svn.metaprl.org/svnroot/mojave/omake-branches/0.9.8.x@10130 5109f7fa-8300-0410-942d-a7a480986ffa