Further optimized the path code, finally bringing the performance close to what it was before the hash-cons patch was applied.
The problem was that when appending an extra path (which might contain ".."
somehere inside) to a directory, it would fully reconstruct the directory and
then construct it from scratch again. In presence of cons-hashing this is
really expensive.

The obvious solution is not to deconstruct and only pop one level every time
the ".." is encountered.


git-svn-id: http://svn.metaprl.org/svnroot/mojave/omake-branches/0.9.8.x@10258 5109f7fa-8300-0410-942d-a7a480986ffa