A few more declarations.
Minor: when choosing names, please try not to make silly abbreviations,
like EMPTY_ARR.  What is an ARR?  For two characters more, it becomes
very clear, EMPTY_ARRAY.  The other problem with abbreviations is that
you have to remember which names are abbreviated, and which are not.
It is a lot easier to be consistent.  The standard abbreviations are
acceptable of course, like int, float, dir, etc.  In Pervasives.om
at least, spell it out.


git-svn-id: http://svn.metaprl.org/svnroot/mojave/omake-branches/0.9.8.x@10482 5109f7fa-8300-0410-942d-a7a480986ffa