Fixed an uncaught exception in the case-sensitive test.
This was a little exasperating--I was wondering why all of a sudden
I was getting End_of_file exceptions in the middle of a file.

I believe it is more important for the code to be clean and correct
than to have complicated incorrect code that is faster by a few
microseconds.  After all, these micro-optimizations are really
swamped by the system calls.  Unix.LargeFile.stat is very expensive--the
only way you are going to make this code faster in any significant
way is by reducing the number of system calls.


git-svn-id: http://svn.metaprl.org/svnroot/mojave/omake-branches/0.9.8.x@10555 5109f7fa-8300-0410-942d-a7a480986ffa