Added the array arguments, bug #645.  The syntax is:    - ... for a normal body argument    - [...] for an array body argument
   X =
      file([...])
         a
         b
         c
   - : <array
       /Users/jyh/projects/omake/0.9.8.x/a : File
       /Users/jyh/projects/omake/0.9.8.x/b : File
       /Users/jyh/projects/omake/0.9.8.x/c : File>
       : Array

If an application has an elision, you can omit the colon
at the end in interactive mode.

This can still be improved, especially with keyword
arguments.  When there are keyword arguments, the argument
lines can get long.  It might be nice to allow using
a line break instead of the , separator.

   f(X[] =
        a
        b
        c
     foo
     Y = 1)

Or perhaps allow mixed notation.

   f(X = 1
     Y = 2
     a, b, c)


git-svn-id: http://svn.metaprl.org/svnroot/mojave/omake-branches/0.9.8.x@10613 5109f7fa-8300-0410-942d-a7a480986ffa