Here is a proposal for fixing the case-sensitive problem.
The basic idea is that we should enforce case-sensitivity
even on insensitive filesystems.  If that is too strong, then
we should at least enforce it during scanning.  The reason is
that when you port your case-insensitive project to some
other system, you want the build to work the same.

We can't just enforce case-sensitivity by turning off the detection.
The proper solution requires figuring out the actual filenames.

This commit adds Node.real_tail and related functions that
returns the actual filename for the node, which it uses to
implement a strict (case-sensitive) form of target-is-buildable.  Let
me know what you think.


git-svn-id: http://svn.metaprl.org/svnroot/mojave/omake-branches/0.9.8.x@10637 5109f7fa-8300-0410-942d-a7a480986ffa