Our current algorithm for the "generic" case-sensitivity test was:     1. Try the name being created itself.     2. Try looking at the directory entries.     3. Create a dummy file, and test.     4. Test the parent.
Step 2 would start with the Unix.opendir call and if that failed we would
abvance to step 3. 

This commit changes it so that if the Unix.opendir error code tells us the
directory does not exist (either of: ENOENT, ENOTDIR, ELOOP, or ENAMETOOLONG),
then we advance to step 4 right away.

Note that in the case of a permission error in Unix.opendir, we still go to
step 3 (as a directory with "wx" permission, but no "r" would allow step 3,
but not opendir).


git-svn-id: http://svn.metaprl.org/svnroot/mojave/omake-branches/0.9.8.x@10861 5109f7fa-8300-0410-942d-a7a480986ffa