Implement the --configure for the .STATIC sections correctly. Previous implementation would simply fail to consult the Omake_cache when --configure is given. However, the Omake_cache is used both for the caching of the .STATIC sections to disk and for the "memoization" of the .STATIC sections, while only the former function is supposed to be affected by --configure.
The new implementation only checks for the --configure option when the cache
is loaded from disk - and drops the loaded value table if --configure is
present.

This commit also pushes the processing of all the cache-related options
(including -U and --depend) into Omake_cache.from_channel, instead of having
each caller of Omake_cache.from_channel (currently we have two) have to do it
independently.


git-svn-id: http://svn.metaprl.org/svnroot/mojave/omake-branches/0.9.8.x@11185 5109f7fa-8300-0410-942d-a7a480986ffa