The hoisting Aleksey added in revs 11448-11849 was not right-- the following expression doesn't make sense:
   venv_export_venv venv_orig venv_src

This is because the two environments have different this.
venv_orig has the caller's "this", venv_src has the callee's "this".
The two objects are different, so the export makes no sense.

NOTE: disabled the path check, going back to the original policy
that hoisting overrides export.  This is matter of policy, so I
think it is better to keep the code simple, and discuss it.

Also, changed the tests a little, so an exit code of 2
is a "note", not an "error".


git-svn-id: http://svn.metaprl.org/svnroot/mojave/omake-branches/0.9.8.x@11850 5109f7fa-8300-0410-942d-a7a480986ffa