Initial reorganization of doc/html. What used to be doc/html is now split into:
- doc/webpage - manually edited web files (moved from doc/html)
- doc/webpage/images - gifs for the web page (copied from doc/html/images)
- doc/webpage/manual - auto-generated files _snapshot_ (copied from doc/html)
- doc/html - latest auto-generated files (all manual files deleted)
- doc/images - all image stuff - .gif, .ai, .eps, etc (renamed from
  doc/html/images)

The idea is to point the web page at doc/webpage, where the doc/webpage/manual
would only ever contain "release-worthy" snapshots of the auto-generated files
(even if it is on some obscure branch). The doc/html would now be a separate
repository for the latest generated doc files, never directly exposed to the
main web page). Posting the latest documentation on the web would now require
one extra step (copying from doc/html to doc/webpage/manual and from
doc/images/*{gif,jpg,png} to doc/webpage/images) - essentially "blessing" a
particular auto-generated version, so I think it's a good thing to have it.

We can later try to follow-up with Jason's idea of putting all the sources
into doc/src, although it may be a bit tricky (as we'd want the relative links
to work for every location of each file).

After this reorg, the following files are "primary" (and can be edited
safely):
- doc/src/*
- doc/images/*
- doc/webpage/* (but _not_ doc/webpage/*/*)



git-svn-id: http://svn.metaprl.org/svnroot/mojave/omake-branches/0.9.8.x@12426 5109f7fa-8300-0410-942d-a7a480986ffa