Removing the Om_symbol module, restoring the usage of Lm_symbol, which is a symlink to Lm_symbol_hash implelemtation.
This is, essentially, a reversal of rev. 12585

P.S. We may want to revisit this after the "jumbo" stuff is fully landed, but
currently it's too big of a hassle to try to push these changes through the
branches. Also it's not clear why Om_symbol might be better then
Lm_symbol_hash (the two implementations are almost identical).


git-svn-id: http://svn.metaprl.org/svnroot/mojave/omake-branches/0.9.8.x@12664 5109f7fa-8300-0410-942d-a7a480986ffa